requirements.txt
Flask
numpy
Pillow
selenium
webdriver-manager